Separate attestation processing into a separate class.
