+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/SpecLogic.java
@@ -19,6 +19,7 @@ import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateAccessors;
 import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateMutators;
 import tech.pegasys.teku.spec.logic.common.helpers.MiscHelpers;
 import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
+import tech.pegasys.teku.spec.logic.common.operations.attestation.AttestationProcessor;
 import tech.pegasys.teku.spec.logic.common.statetransition.StateTransition;
 import tech.pegasys.teku.spec.logic.common.statetransition.epoch.EpochProcessor;
 import tech.pegasys.teku.spec.logic.common.statetransition.epoch.status.ValidatorStatusFactory;
@@ -42,6 +43,8 @@ public interface SpecLogic {
   EpochProcessor getEpochProcessor();
 
   BlockProcessor getBlockProcessor();
+
+  AttestationProcessor getAttestationProcessor();
 
   StateTransition getStateTransition();
