+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/altair/block/BlockProcessorAltair.java
@@ -16,7 +16,6 @@ package tech.pegasys.teku.spec.logic.versions.altair.block;
 import static tech.pegasys.teku.spec.constants.IncentivizationWeights.PROPOSER_WEIGHT;
 import static tech.pegasys.teku.spec.constants.IncentivizationWeights.SYNC_REWARD_WEIGHT;
 import static tech.pegasys.teku.spec.constants.IncentivizationWeights.WEIGHT_DENOMINATOR;
-import static tech.pegasys.teku.spec.logic.common.helpers.MathHelpers.integerSquareRoot;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,35 +25,27 @@ import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.bls.BLSSignature;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfigAltair;
-import tech.pegasys.teku.spec.constants.ParticipationFlags;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.SyncAggregate;
-import tech.pegasys.teku.spec.datastructures.operations.Attestation;
-import tech.pegasys.teku.spec.datastructures.operations.AttestationData;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
-import tech.pegasys.teku.spec.datastructures.state.Checkpoint;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.altair.MutableBeaconStateAltair;
 import tech.pegasys.teku.spec.datastructures.type.SszPublicKey;
 import tech.pegasys.teku.spec.logic.common.block.AbstractBlockProcessor;
 import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateMutators;
 import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
-import tech.pegasys.teku.spec.logic.common.operations.validation.AttestationDataStateTransitionValidator;
+import tech.pegasys.teku.spec.logic.common.operations.attestation.AttestationProcessor;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.BlockProcessingException;
 import tech.pegasys.teku.spec.logic.common.util.AttestationUtil;
 import tech.pegasys.teku.spec.logic.common.util.BeaconStateUtil;
 import tech.pegasys.teku.spec.logic.common.util.ValidatorsUtil;
 import tech.pegasys.teku.spec.logic.versions.altair.helpers.BeaconStateAccessorsAltair;
 import tech.pegasys.teku.spec.logic.versions.altair.helpers.MiscHelpersAltair;
-import tech.pegasys.teku.spec.logic.versions.altair.helpers.MiscHelpersAltair.FlagIndexAndWeight;
-import tech.pegasys.teku.ssz.SszMutableList;
 import tech.pegasys.teku.ssz.SszVector;
-import tech.pegasys.teku.ssz.collections.SszUInt64List;
 import tech.pegasys.teku.ssz.primitive.SszByte;
 import tech.pegasys.teku.ssz.primitive.SszUInt64;
 
 public class BlockProcessorAltair extends AbstractBlockProcessor {
   private final SpecConfigAltair specConfigAltair;
-  private final MiscHelpersAltair miscHelpersAltair;
   private final BeaconStateAccessorsAltair beaconStateAccessorsAltair;
 
   public BlockProcessorAltair(
@@ -66,7 +57,7 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
       final BeaconStateUtil beaconStateUtil,
       final AttestationUtil attestationUtil,
       final ValidatorsUtil validatorsUtil,
-      final AttestationDataStateTransitionValidator attestationValidator) {
+      final AttestationProcessor attestationProcessor) {
     super(
         specConfig,
         predicates,
@@ -76,96 +67,10 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
         beaconStateUtil,
         attestationUtil,
         validatorsUtil,
-        attestationValidator);
+        attestationProcessor);
 
     this.specConfigAltair = specConfig;
-    this.miscHelpersAltair = miscHelpers;
     this.beaconStateAccessorsAltair = beaconStateAccessors;
-  }
-
-  @Override
-  protected void processAttestation(
-      final MutableBeaconState genericState,
-      final Attestation attestation,
-      final IndexedAttestationProvider indexedAttestationProvider) {
-    final MutableBeaconStateAltair state = MutableBeaconStateAltair.required(genericState);
-    final AttestationData data = attestation.getData();
-
-    final SszMutableList<SszByte> epochParticipation;
-    final Checkpoint justifiedCheckpoint;
-    if (data.getTarget().getEpoch().equals(beaconStateAccessors.getCurrentEpoch(state))) {
-      epochParticipation = state.getCurrentEpochParticipation();
-      justifiedCheckpoint = state.getCurrent_justified_checkpoint();
-    } else {
-      epochParticipation = state.getPreviousEpochParticipation();
-      justifiedCheckpoint = state.getPrevious_justified_checkpoint();
-    }
-
-    // Matching roots
-    final boolean isMatchingHead =
-        data.getBeacon_block_root()
-            .equals(beaconStateUtil.getBlockRootAtSlot(state, data.getSlot()));
-    final boolean isMatchingSource = data.getSource().equals(justifiedCheckpoint);
-    final boolean isMatchingTarget =
-        data.getTarget()
-            .getRoot()
-            .equals(beaconStateUtil.getBlockRoot(state, data.getTarget().getEpoch()));
-
-    // Participation flag indices
-    final List<Integer> participationFlagIndices = new ArrayList<>();
-    final UInt64 stateSlot = state.getSlot();
-    final UInt64 dataSlot = data.getSlot();
-    if (isMatchingHead
-        && isMatchingTarget
-        && stateSlot.equals(dataSlot.plus(specConfig.getMinAttestationInclusionDelay()))) {
-      participationFlagIndices.add(ParticipationFlags.TIMELY_HEAD_FLAG_INDEX);
-    }
-    if (isMatchingSource
-        && stateSlot.isLessThanOrEqualTo(
-            dataSlot.plus(integerSquareRoot(specConfig.getSlotsPerEpoch())))) {
-      participationFlagIndices.add(ParticipationFlags.TIMELY_SOURCE_FLAG_INDEX);
-    }
-    if (isMatchingTarget
-        && stateSlot.isLessThanOrEqualTo(dataSlot.plus(specConfig.getSlotsPerEpoch()))) {
-      participationFlagIndices.add(ParticipationFlags.TIMELY_TARGET_FLAG_INDEX);
-    }
-
-    // Update epoch participation flags
-    UInt64 proposerRewardNumerator = UInt64.ZERO;
-    final SszUInt64List attestingIndices =
-        indexedAttestationProvider.getIndexedAttestation(attestation).getAttesting_indices();
-    for (SszUInt64 attestingIndex : attestingIndices) {
-      final int index = attestingIndex.get().intValue();
-      byte participationFlags = epochParticipation.get(index).get();
-      final UInt64 baseReward = beaconStateAccessorsAltair.getBaseReward(state, index);
-      boolean shouldUpdate = false;
-      for (FlagIndexAndWeight flagIndicesAndWeight : miscHelpersAltair.getFlagIndicesAndWeights()) {
-        final int flagIndex = flagIndicesAndWeight.getIndex();
-        final UInt64 weight = flagIndicesAndWeight.getWeight();
-
-        if (participationFlagIndices.contains(flagIndex)
-            && !miscHelpersAltair.hasFlag(participationFlags, flagIndex)) {
-          shouldUpdate = true;
-          participationFlags = miscHelpersAltair.addFlag(participationFlags, flagIndex);
-          proposerRewardNumerator = proposerRewardNumerator.plus(baseReward.times(weight));
-        }
-      }
-
-      if (shouldUpdate) {
-        epochParticipation.set(index, SszByte.of(participationFlags));
-      }
-    }
-
-    if (!proposerRewardNumerator.isZero()) {
-      final int proposerIndex = beaconStateAccessors.getBeaconProposerIndex(state);
-      final UInt64 proposerRewardDenominator =
-          WEIGHT_DENOMINATOR
-              .minus(PROPOSER_WEIGHT)
-              .times(WEIGHT_DENOMINATOR)
-              .dividedBy(PROPOSER_WEIGHT);
-      final UInt64 proposerReward = proposerRewardNumerator.dividedBy(proposerRewardDenominator);
-      beaconStateMutators.increaseBalance(state, proposerIndex, proposerReward);
-    }
   }
 
   @Override
